move to use Expression node to check Throw inside Expression
