+++ rules-tests/DowngradePhp80/Rector/Expression/DowngradeThrowExprRector/config/configured_rule.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-use Rector\DowngradePhp80\Rector\Throw_\DowngradeThrowExprRector;
+use Rector\DowngradePhp80\Rector\Expression\DowngradeThrowExprRector;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
 return static function (ContainerConfigurator $containerConfigurator): void {
